+++ src/shared_utilities/models.jl
@@ -21,7 +21,9 @@ export AbstractModel,
     make_update_cache,
     add_drivers_to_cache,
     get_drivers,
-    name
+    name,
+    total_liq_water_vol_per_area!,
+    total_energy_per_area!
 
 import ClimaComms
 import .Domains: coordinates
@@ -466,3 +468,28 @@ function ClimaComms.device(model::AbstractModel)
         )
     end
 end
+
+"""
+    total_liq_water_vol_per_area!(cache, model::AbstractModel, Y, p, t)
+
+A function which updates `cache` in place with the total liquid water volume
+per unit ground area for the `model`, computed from `Y`, `p`, and `t`.
+
+While water mass is the fundamentally conserved quantity, soil modelling represents
+water by an equivalent water volume using the density of water and ice at standard
+temperature and pressure. Because of that, we report here the total volume of water
+present in a model (per unit area) that would arise if all the water was in liquid phase.
+This can be converted to a mass using the density of liquid water.
+
+This includes the water in multiple phases. For example, if ice is present, the water
+volume is computed using ratio of the density of ice to the density of liquid water.
+"""
+function total_liq_water_vol_per_area!(cache, model::AbstractModel, Y, p, t) end
+
+"""
+    total_energy_per_area!(cache, model::AbstractModel, Y, p, t)
+
+A function which updates `cache` in place with the total energy
+per unit ground area for the `model`, computed from `Y`, `p`, and `t`.
+"""
+function total_energy_per_area!(cache, model::AbstractModel, Y, p, t) end